Vereinfachung Produktion
Ab sofort reicht es, im Konstruktor des jeweiligen Produktionsgebäude
die possible-Liste um die ProductTypes zu ergänzen. possible wird
bereits in PorductionBuilding initialisiert und auch dort verwendet.
